+++ Source/Interval.cpp
@@ -4,19 +4,19 @@
 namespace Pooraytracer {
 
 	const Interval Interval::empty = Interval(
-		+std::numeric_limits<float>::infinity(),
-		-std::numeric_limits<float>::infinity()
+		+std::numeric_limits<double>::infinity(),
+		-std::numeric_limits<double>::infinity()
 	);
 	const Interval Interval::universe = Interval(
-		-std::numeric_limits<float>::infinity(),
-		+std::numeric_limits<float>::infinity()
+		-std::numeric_limits<double>::infinity(),
+		+std::numeric_limits<double>::infinity()
 	);
 
-	Interval operator+(const Interval& ival, float displacement)
+	Interval operator+(const Interval& ival, double displacement)
 	{
 		return Interval(ival.min + displacement, ival.max + displacement);
 	}
-	Interval operator+(float displacement, const Interval& ival)
+	Interval operator+(double displacement, const Interval& ival)
 	{
 		return ival + displacement;
 	}
