+++ Source/Interval.h
@@ -7,37 +7,37 @@ namespace Pooraytracer {
 
 	class Interval {
 	public:
-		float min, max;
-		Interval() :min(+std::numeric_limits<float>::infinity()), max(-std::numeric_limits<float>::infinity()) {}
-		Interval(float min, float max) : min(min), max(max) {}
+		double min, max;
+		Interval() :min(+std::numeric_limits<double>::infinity()), max(-std::numeric_limits<double>::infinity()) {}
+		Interval(double min, double max) : min(min), max(max) {}
 		Interval(const Interval& a, const Interval& b) {
 			min = a.min <= b.min ? a.min : b.min;
 			max = a.max >= b.max ? a.max : b.max;
 		}
 
-		float Length() const {
+		double Length() const {
 			return max - min;
 		}
-		bool Contains(float x) const {
+		bool Contains(double x) const {
 			return min <= x && x <= max;
 		}
-		bool Surrounds(float x) const {
+		bool Surrounds(double x) const {
 			return min < x && x < max;
 		}
-		float Clamp(float x) const {
+		double Clamp(double x) const {
 			if (x < min) return min;
 			if (x > max) return max;
 			return x;
 		}
 
-		Interval Expand(float delta) const {
-			float padding = delta / 2.f;
+		Interval Expand(double delta) const {
+			double padding = delta / 2.;
 			return Interval(min - padding, max + padding);
 		}
 
 		static const Interval empty, universe;
 	};
 
-	Interval operator+(const Interval& ival, float displacement);
-	Interval operator+(float displacement, const Interval& ival);
+	Interval operator+(const Interval& ival, double displacement);
+	Interval operator+(double displacement, const Interval& ival);
 }
